2011-01-03  Simon Fraser  <simon.fraser@apple.com>
        Fix warning in 32-bit builds.

        * css/CSSGradientValue.cpp:
        (WebCore::CSSLinearGradientValue::createGradient):

Canonical link: https://commits.webkit.org/65258@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@74919 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/css/CSSGradientValue.cpp
@@ -440,7 +440,7 @@ PassRefPtr<Gradient> CSSLinearGradientValue::createGradient(RenderObject* render
     FloatPoint firstPoint;
     FloatPoint secondPoint;
     if (m_angle) {
-        float angle = m_angle->getDoubleValue(CSSPrimitiveValue::CSS_DEG);
+        float angle = m_angle->getFloatValue(CSSPrimitiveValue::CSS_DEG);
         endPointsFromAngle(angle, size, firstPoint, secondPoint);
     } else {
         firstPoint = computeEndPoint(m_firstX.get(), m_firstY.get(), renderer->style(), rootStyle, size);
